+++ src/index.css
@@ -21,7 +21,7 @@
   --golden: #f5e689;
   --green: #4da406;
   --purple: #13022d;
-  --purple-background: #0b0118e7;
+  --purple-background: #0b0118cb;
   --grey-text: #5b57573d;
 }
