+++ src/pages/Home.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import styled from "styled-components";
 import Featured from "../components/organisms/Featured/Featured";
 import Navbar from "../components/molecules/Navbar";
@@ -12,11 +12,32 @@ import star from "../assets/bubble.jfif";
 const StyledMain = styled.main`
   position: relative;
   padding: 0 3rem;
+  width: 100vw;
 `;
+
+function debounce(fn, ms) {
+  let timer;
+  return () => {
+    clearTimeout(timer);
+    timer = setTimeout(fn, ms);
+  };
+}
+
 
 const Home = () => {
   const [theme, setTheme] = useState(true);
+  const [screen, setScreen] = useState(window.innerWidth);
+  useEffect(() => {
+    const debouncedHandleResize = debounce(function resize() {
+      setScreen(window.innerWidth);
+    }, 1000);
 
+    window.addEventListener("resize", debouncedHandleResize);
+
+    return () => {
+      window.removeEventListener("resize", debouncedHandleResize);
+    };
+  }, []);
   return (
     <StyledMain
       theme={theme}
@@ -31,10 +52,10 @@ const Home = () => {
       <SocialMedia />
       <Navbar setTheme={setTheme} theme={theme} />
       <Container>
-        <Header />
+        <Header screen={screen} />
       </Container>
       <Container>
-        <AboutMe />
+        <AboutMe screen={screen} />
       </Container>
       <Container>
         <Featured />
